+++ changelog.md
@@ -1,11 +1,28 @@
 # Changelog for PSFunctionTools
 
-## v1.2.0
+## [Unreleased]
+
+## [1.3.0] - 2025-01-14
+
+### Added
+
+- Added command `Open-PSFunctionToolsSamples` to change to the Samples folder and list the sample files.
+- Added command `Open-PSFunctionToolsHelp`.
+
+### Changed
+
+- Updated demo file `Demo-NewModuleFromFiles.ps1`.
+- Help documentation updates.
+- Minor code cleanup.
+- Bumped minimum PowerShell version to 7.4. __This is a potential breaking change.__
+- Converted changelog to new format.
+
+## [v1.2.0] - 2023-04-18
 
 ### Changed
 
 - Pester test revisions.
-- Modified `Format-FunctionName` to let you capitalize N number of characters in the Noun portion of your command name.
+- Modified `Format-FunctionName` to let you capitalize N number of characters in the Noun portion of the command name.
 - Help updates.
 - Updated `README.md`.
 
@@ -17,14 +34,14 @@
 
 - Deleted unused `Types` folder.
 
-## v1.1.0
+## [v1.1.0] - 2023-03-03
 
 ### Changed
 
 - Code clean up and reformatting.
 - Image resizing.
 
-## v1.0.0
+## [v1.0.0] - 2022-02-28
 
 - Added online help links for previously published commands.
 - Bumped version number to reflect feature-complete release.
@@ -32,7 +49,7 @@
 - Help updates.
 - Updated `README.md`.
 
-## v0.6.0
+## [v0.6.0] - 2022-01-17
 
 - Added online help links for previously published commands.
 - Modified `Get-FunctionAttribute` to accept pipeline input for `Name` and `Path` parameters.
@@ -47,7 +64,7 @@
 - Help updates.
 - Initial public release to the PowerShell Gallery.
 
-## v0.5.0
+## v0.5.0 - 2022-01-11
 
 - Module will be written to support PowerShell 7.1 and later. Commands may run in Windows PowerShell but I am __not__ marking the module as `Desktop` compatible.
 - Added a parameter called `AsTree` to `Get-ModuleLayout` to show module layout as a tree.
@@ -58,14 +75,14 @@
 - Updated `README.md`.
 - First public preview release.
 
-## v0.4.0
+## v0.4.0 - 2021-12-24
 
 - Moved helper functions in `New-ModuleFromFiles` to `functions\private\helpers.ps1'`.
-- Updated `New-ModuleFromFiles` to use `Functionpath` parameter when creating the root psm1 file.
+- Updated `New-ModuleFromFiles` to use `FunctionPath` parameter when creating the root psm1 file.
 - Added function `New-ModuleFromLayout`.
 - Help updates.
 
-## v0.3.0
+## v0.3.0 - 2021-12-21
 
 - Added sample scripts.
 - Modified `New-CommentHelp` with a `-TemplateOnly` parameter to generate help without any parameter definitions.
@@ -73,7 +90,7 @@
 - Modified `Get-ParameterBlock` to accept `Path` and `Name` values from pipeline input by property name.
 - Added function `New-ModuleFromFiles`. This should be considered experimental.
 
-## v0.2.0
+## v0.2.0 - 2021-12-20
 
 - Updated help documentation.
 - Modified `Format-FunctionName` to accept pipeline input by value.
@@ -81,9 +98,20 @@
 - Added function `Get-FunctionAttribute` with an alias of `gfa`.
 - Added a parameter called `Detailed` to `Get-FunctionName` to write a custom object to the pipeline which includes the path to the file. Added a custom format file `psfunctionname.format.ps1xml`.
 
-## v0.1.0
+## v0.1.0 - 2021-12-20
 
 - Initial files.
 - Added `Get-ModuleLayout` with format file `modulelayout.format.ps1xml`.
 - Modified `Get-PSScriptRequirements` to write a `PSScriptRequirements` object to the pipeline. Added the format file `psscriptrequirements.format.ps1xml`.
 - Updated help documentation.
+
+[Unreleased]: https://github.com/jdhitsolutions/PSFunctionTools/compare/v1.3.0..HEAD
+[1.3.0]: https://github.com/jdhitsolutions/PSFunctionTools/compare/vv1.2.0..v1.3.0
+[v1.2.0]: https://github.com/jdhitsolutions/PSFunctionTools/compare/v1.1.0..v1.2.0
+[v1.1.0]: https://github.com/jdhitsolutions/PSFunctionTools/compare/v1.0.0..v1.1.0
+[v1.0.0]: https://github.com/jdhitsolutions/PSFunctionTools/compare/v0.6.0..v1.0.0
+[v0.6.0]: https://github.com/jdhitsolutions/PSFunctionTools/compare/v0.5.0..v0.6.0
+[v0.5.0]:
+[v0.4.0]:
+[v0.3.0]:
+[v0.2.0]: